separate the cli file and the class file
